+++ proxmox/virtual_environment_nodes_types.go
@@ -17,6 +17,17 @@ type VirtualEnvironmentNodeExecuteRequestBody struct {
 	Commands CustomNodeCommands `json:"commands" url:"commands"`
 }
 
+// VirtualEnvironmentNodeGetTaskStatusResponseBody contains the body from a node get task status response.
+type VirtualEnvironmentNodeGetTaskStatusResponseBody struct {
+	Data *VirtualEnvironmentNodeGetTaskStatusResponseData `json:"data,omitempty"`
+}
+
+// VirtualEnvironmentNodeGetTaskStatusResponseData contains the data from a node get task status response.
+type VirtualEnvironmentNodeGetTaskStatusResponseData struct {
+	PID    int    `json:"pid,omitempty"`
+	Status string `json:"status,omitempty"`
+}
+
 // VirtualEnvironmentNodeListResponseBody contains the body from a node list response.
 type VirtualEnvironmentNodeListResponseBody struct {
 	Data []*VirtualEnvironmentNodeListResponseData `json:"data,omitempty"`