+++ proxmoxtf/resource_virtual_environment_vm.go
@@ -433,7 +433,6 @@ func resourceVirtualEnvironmentVM() *schema.Resource {
 							Type:        schema.TypeString,
 							Description: "The datastore id",
 							Optional:    true,
-							ForceNew:    true,
 							Default:     dvResourceVirtualEnvironmentVMDiskDatastoreID,
 						},
 						mkResourceVirtualEnvironmentVMDiskFileFormat: {
@@ -456,7 +455,6 @@ func resourceVirtualEnvironmentVM() *schema.Resource {
 							Type:         schema.TypeInt,
 							Description:  "The disk size in gigabytes",
 							Optional:     true,
-							ForceNew:     true,
 							Default:      dvResourceVirtualEnvironmentVMDiskSize,
 							ValidateFunc: validation.IntBetween(1, 8192),
 						},
@@ -1668,13 +1666,13 @@ func resourceVirtualEnvironmentVMCreateStart(d *schema.ResourceData, m interface
 	}
 
 	// Start the virtual machine and wait for it to reach a running state before continuing.
-	err = veClient.StartVM(nodeName, vmID)
+	taskID, err := veClient.StartVM(nodeName, vmID)
 
 	if err != nil {
 		return err
 	}
 
-	err = veClient.WaitForVMState(nodeName, vmID, "running", 120, 5)
+	err = veClient.WaitForNodeTask(nodeName, *taskID, 600, 5)
 
 	if err != nil {
 		return err
@@ -3274,13 +3272,13 @@ func resourceVirtualEnvironmentVMUpdate(d *schema.ResourceData, m interface{}) e
 
 	if d.HasChange(mkResourceVirtualEnvironmentVMStarted) && !bool(template) {
 		if started {
-			err = veClient.StartVM(nodeName, vmID)
+			taskID, err := veClient.StartVM(nodeName, vmID)
 
 			if err != nil {
 				return err
 			}
 
-			err = veClient.WaitForVMState(nodeName, vmID, "running", 120, 5)
+			err = veClient.WaitForNodeTask(nodeName, *taskID, 600, 5)
 
 			if err != nil {
 				return err
@@ -3289,7 +3287,7 @@ func resourceVirtualEnvironmentVMUpdate(d *schema.ResourceData, m interface{}) e
 			forceStop := proxmox.CustomBool(true)
 			shutdownTimeout := 300
 
-			err = veClient.ShutdownVM(nodeName, vmID, &proxmox.VirtualEnvironmentVMShutdownRequestBody{
+			taskID, err := veClient.ShutdownVM(nodeName, vmID, &proxmox.VirtualEnvironmentVMShutdownRequestBody{
 				ForceStop: &forceStop,
 				Timeout:   &shutdownTimeout,
 			})
@@ -3298,7 +3296,7 @@ func resourceVirtualEnvironmentVMUpdate(d *schema.ResourceData, m interface{}) e
 				return err
 			}
 
-			err = veClient.WaitForVMState(nodeName, vmID, "stopped", 30, 5)
+			err = veClient.WaitForNodeTask(nodeName, *taskID, 600, 5)
 
 			if err != nil {
 				return err
@@ -3308,11 +3306,131 @@ func resourceVirtualEnvironmentVMUpdate(d *schema.ResourceData, m interface{}) e
 		}
 	}
 
-	// Reboot the virtual machine, if required.
-	if !bool(template) && rebootRequired {
+	// Change the disk locations and/or sizes, if necessary.
+	return resourceVirtualEnvironmentVMUpdateDiskLocationAndSize(d, m, vmConfig, !bool(template) && rebootRequired)
+}
+
+func resourceVirtualEnvironmentVMUpdateDiskLocationAndSize(d *schema.ResourceData, m interface{}, vmConfig *proxmox.VirtualEnvironmentVMGetResponseData, reboot bool) error {
+	config := m.(providerConfiguration)
+	veClient, err := config.GetVEClient()
+
+	if err != nil {
+		return err
+	}
+
+	nodeName := d.Get(mkResourceVirtualEnvironmentVMNodeName).(string)
+	started := d.Get(mkResourceVirtualEnvironmentVMStarted).(bool)
+	template := d.Get(mkResourceVirtualEnvironmentVMTemplate).(bool)
+	vmID, err := strconv.Atoi(d.Id())
+
+	if err != nil {
+		return err
+	}
+
+	// Determine if any of the disks are changing location and/or size, and initiate the necessary actions.
+	if d.HasChange(mkResourceVirtualEnvironmentVMDisk) {
+		diskOld, diskNew := d.GetChange(mkResourceVirtualEnvironmentVMDisk)
+
+		diskOldEntries := diskOld.([]interface{})
+		diskNewEntries := diskNew.([]interface{})
+
+		diskMoveBodies := []*proxmox.VirtualEnvironmentVMMoveDiskRequestBody{}
+		diskResizeBodies := []*proxmox.VirtualEnvironmentVMResizeDiskRequestBody{}
+
+		for i := range diskOldEntries {
+			diskOldBlock := diskOldEntries[i].(map[string]interface{})
+			diskNewBlock := diskNewEntries[i].(map[string]interface{})
+
+			diskOldDatastoreID := diskOldBlock[mkResourceVirtualEnvironmentVMDiskDatastoreID].(string)
+			diskNewDatastoreID := diskNewBlock[mkResourceVirtualEnvironmentVMDiskDatastoreID].(string)
+
+			if diskOldDatastoreID != diskNewDatastoreID {
+				deleteOriginalDisk := proxmox.CustomBool(true)
+
+				diskMoveBodies = append(diskMoveBodies, &proxmox.VirtualEnvironmentVMMoveDiskRequestBody{
+					DeleteOriginalDisk: &deleteOriginalDisk,
+					Disk:               fmt.Sprintf("scsi%d", i),
+					TargetStorage:      diskNewDatastoreID,
+				})
+			}
+
+			diskOldSize := diskOldBlock[mkResourceVirtualEnvironmentVMDiskSize].(int)
+			diskNewSize := diskNewBlock[mkResourceVirtualEnvironmentVMDiskSize].(int)
+
+			if diskOldSize != diskNewSize {
+				diskResizeBodies = append(diskResizeBodies, &proxmox.VirtualEnvironmentVMResizeDiskRequestBody{
+					Disk: fmt.Sprintf("scsi%d", i),
+					Size: fmt.Sprintf("%dG", diskNewSize),
+				})
+			}
+		}
+
+		if len(diskMoveBodies) > 0 || len(diskResizeBodies) > 0 {
+			if !template {
+				forceStop := proxmox.CustomBool(true)
+				shutdownTimeout := 300
+
+				taskID, err := veClient.ShutdownVM(nodeName, vmID, &proxmox.VirtualEnvironmentVMShutdownRequestBody{
+					ForceStop: &forceStop,
+					Timeout:   &shutdownTimeout,
+				})
+
+				if err != nil {
+					return err
+				}
+
+				err = veClient.WaitForNodeTask(nodeName, *taskID, 600, 5)
+
+				if err != nil {
+					return err
+				}
+			}
+
+			reboot = false
+		}
+
+		for _, reqBody := range diskMoveBodies {
+			taskID, err := veClient.MoveVMDisk(nodeName, vmID, reqBody)
+
+			if err != nil {
+				return err
+			}
+
+			err = veClient.WaitForNodeTask(nodeName, *taskID, 86400, 5)
+
+			if err != nil {
+				return err
+			}
+		}
+
+		for _, reqBody := range diskResizeBodies {
+			err := veClient.ResizeVMDisk(nodeName, vmID, reqBody)
+
+			if err != nil {
+				return err
+			}
+		}
+
+		if (len(diskMoveBodies) > 0 || len(diskResizeBodies) > 0) && started && !template {
+			taskID, err := veClient.StartVM(nodeName, vmID)
+
+			if err != nil {
+				return err
+			}
+
+			err = veClient.WaitForNodeTask(nodeName, *taskID, 600, 5)
+
+			if err != nil {
+				return err
+			}
+		}
+	}
+
+	// Perform a regular reboot in case it's necessary and haven't already been done.
+	if reboot {
 		rebootTimeout := 300
 
-		err = veClient.RebootVM(nodeName, vmID, &proxmox.VirtualEnvironmentVMRebootRequestBody{
+		taskID, err := veClient.RebootVM(nodeName, vmID, &proxmox.VirtualEnvironmentVMRebootRequestBody{
 			Timeout: &rebootTimeout,
 		})
 
@@ -3320,13 +3438,10 @@ func resourceVirtualEnvironmentVMUpdate(d *schema.ResourceData, m interface{}) e
 			return err
 		}
 
-		// Wait for the agent to unpublish the network interfaces, if it's enabled.
-		if vmConfig.Agent != nil && vmConfig.Agent.Enabled != nil && *vmConfig.Agent.Enabled {
-			err = veClient.WaitForNoNetworkInterfacesFromVMAgent(nodeName, vmID, 300, 5)
+		err = veClient.WaitForNodeTask(nodeName, *taskID, 600, 5)
 
-			if err != nil {
-				return err
-			}
+		if err != nil {
+			return err
 		}
 	}
 
@@ -3359,7 +3474,7 @@ func resourceVirtualEnvironmentVMDelete(d *schema.ResourceData, m interface{}) e
 		forceStop := proxmox.CustomBool(true)
 		shutdownTimeout := 300
 
-		err = veClient.ShutdownVM(nodeName, vmID, &proxmox.VirtualEnvironmentVMShutdownRequestBody{
+		taskID, err := veClient.ShutdownVM(nodeName, vmID, &proxmox.VirtualEnvironmentVMShutdownRequestBody{
 			ForceStop: &forceStop,
 			Timeout:   &shutdownTimeout,
 		})
@@ -3368,7 +3483,7 @@ func resourceVirtualEnvironmentVMDelete(d *schema.ResourceData, m interface{}) e
 			return err
 		}
 
-		err = veClient.WaitForVMState(nodeName, vmID, "stopped", 30, 5)
+		err = veClient.WaitForNodeTask(nodeName, *taskID, 600, 5)
 
 		if err != nil {
 			return err